pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "codeagents"
version = "3.1.1"
description = "Local coding and chat agents for OpenAI-compatible local runtimes."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4",
  "httpx>=0.27",
  "markdownify>=0.11",
  "mcp>=1.8",
  "psutil>=5.9",
  "pydantic>=2",
  "pypdf>=4.0",
  "sqlite-vec>=0.1.6",
  "tiktoken>=0.7",
]

[project.optional-dependencies]
kg = [
  "python-igraph>=0.11",
  "leidenalg>=0.10",
]

[project.scripts]
codeagents = "codeagents.cli:main"
codeagents-mcp = "codeagents.mcp_server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
